feat: implement feedback system with submission, rating, and average calculation
