+++ core/src/db/schema/index.ts
@@ -1,5 +1,5 @@
 export * from "@/zap/ai/db/schema";
 export * from "@/zap/auth/db/schema";
+export * from "@/zap/feedbacks/db/schema";
 export * from "@/zap/pwa/db/schema";
 export * from "@/zap/waitlist/db/schema";
-export * from "@/zap-old/db/schema/feedbacks.sql";
