+++ core/src/rpc/router.ts
@@ -2,12 +2,12 @@ import "server-only";
 
 import { ai } from "@/zap/ai/rpc/procedures";
 import { auth } from "@/zap/auth/rpc/procedures";
+import { feedbacks } from "@/zap/feedbacks/rpc/procedures";
 import { mails } from "@/zap/mails/rpc/procedures";
 import { pwa } from "@/zap/pwa/rpc/procedures";
+import { waitlist } from "@/zap/waitlist/rpc/procedures";
 import { example } from "@/zap-old/rpc/procedures/example.rpc";
-import { feedbacks } from "@/zap-old/rpc/procedures/feedbacks.rpc";
 import { users } from "@/zap-old/rpc/procedures/users.rpc";
-import { waitlist } from "@/zap-old/rpc/procedures/waitlist.rpc";
 
 export const router = {
   ai,
