+++ core/src/zap/feedbacks/components/feedback-dialog.tsx
@@ -2,7 +2,7 @@
 
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useState } from "react";
-import { useForm } from "react-hook-form";
+import { useForm, UseFormReturn } from "react-hook-form";
 
 import {
   Dialog,
@@ -19,12 +19,14 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Textarea } from "@/components/ui/textarea";
+import { cn } from "@/lib/utils";
 import { ZapButton } from "@/zap/components/button";
-import { RatingButtons } from "@/zap/components/sidebar/rating-buttons";
-import { useIsFeedbackSubmitted } from "@/zap/hooks/feedbacks/use-is-feedback-submitted";
-import { useSubmitFeedback } from "@/zap/hooks/feedbacks/use-submit-feedback";
-import { InputFeedbackSchema } from "@/zap/schemas/feedback.schema";
-import type { FeedbackFormValues } from "@/zap/types/feedback.types";
+import {
+  useIsFeedbackSubmitted,
+  useSubmitFeedback,
+} from "@/zap/feedbacks/hooks";
+import { InputFeedbackSchema } from "@/zap/feedbacks/schemas";
+import type { FeedbackFormValues } from "@/zap/feedbacks/types";
 
 interface FeedbackDialogProps {
   open: boolean;
@@ -123,5 +125,47 @@ export function FeedbackDialog({ open, onOpenChange }: FeedbackDialogProps) {
         )}
       </DialogContent>
     </Dialog>
+  );
+}
+
+interface RatingButtonsProps {
+  form: UseFormReturn<FeedbackFormValues>;
+  isSubmitting: boolean;
+  isExistingFeedback: boolean;
+}
+
+export function RatingButtons({
+  form,
+  isSubmitting,
+  isExistingFeedback,
+}: RatingButtonsProps) {
+  const rating = form.watch("rating");
+
+  return (
+    <div className="flex flex-wrap gap-2 pt-1">
+      {Array.from({ length: 11 }, (_, i) => {
+        const isActive = i <= rating;
+
+        return (
+          <ZapButton
+            aria-label={`Rate ${i}`}
+            className={cn(
+              "h-10 w-10 rounded-md p-0 text-sm",
+              "transform transition-all duration-200 ease-in-out",
+              "hover:scale-110 active:scale-110",
+              isActive && "bg-primary text-white",
+            )}
+            disabled={isSubmitting || !!isExistingFeedback}
+            key={`rating-${i}`}
+            onClick={() => form.setValue("rating", i)}
+            size="sm"
+            type="button"
+            variant={isActive ? "default" : "outline"}
+          >
+            {i}
+          </ZapButton>
+        );
+      })}
+    </div>
   );
 }
