+++ core/src/zap/feedbacks/types/index.ts
@@ -1,5 +1,5 @@
 import type { z } from "zod";
 
-import type { InputFeedbackSchema } from "@/zap/schemas/feedback.schema";
+import type { InputFeedbackSchema } from "@/zap/feedbacks/schemas";
 
 export type FeedbackFormValues = z.infer<typeof InputFeedbackSchema>;
